feat: Refactor tournament types and cleanup tournament utils
Removes the `TournamentStage` enum and moves it to the `Match` interface, as it was not being used at the tournament level. Also, cleans up the `TournamentAdmin` and `TournamentViewer` components by removing unused imports and refactoring the `MatchListItem` component. The `utils/tournament.ts` file is also cleaned up by removing the `shuffleArray` and `calculateStandingsForGroup` functions as they are no longer in use.